+++ hardware/PiUART/src/tcplistenertask.h
@@ -14,7 +14,7 @@
 
 class CTCPListenerTask : public CTask {
 public:
-    CTCPListenerTask(CNetSubSystem *pNet, u16 nListenPort, CRingBuf<u8> *pToSerial, TCPMode TMode, CCharConv *pCharConv);
+    CTCPListenerTask(CNetSubSystem *pNet, u16 nListenPort, CRingBuf<u16> *pToSerial, TCPMode TMode, CCharConv *pCharConv);
     ~CTCPListenerTask();
 
     void Run();
@@ -25,7 +25,7 @@ public:
 private:
     CNetSubSystem *m_pNet;
     u16 m_nListenPort;
-    CRingBuf<u8> *m_pToSerial;
+    CRingBuf<u16> *m_pToSerial;
     CSocket *m_pListenSocket;
     CTCPTelnetTask *m_pTelnetTask;
     CTCPRawTask *m_pRawTask;
